fix: :bug: Fix expense page load with current month
This update resolves the issue where the expense page was not displaying the current month upon loading. The root cause was identified as passing a string instead of a number to the default value of the month select filter. Additionally, the month number is now sent in the URL to the category details page for month-based filtering. All console.log statements have been removed from the data file.

Resolved the expense page loading issue by correcting the default month select filter value. Enabled month-based filtering on the category details page by passing the month number in the URL. Cleaned up the code by removing unnecessary console.log statements from the data file.  Fixed the display issue of the current month on the expense page. Enabled more accurate filtering on the category details page. Improved code cleanliness and performance by removing debug statements.

+++ src/app/ui/monedex/filter-month.tsx
@@ -60,8 +60,7 @@ export default function FilterMonth(): JSX.Element {
   const filterParams = useSearchParams()
   const pathname = usePathname()
   const { replace } = useRouter()
-  const [currentMonth, setCurrentMonth] = useState(new Date().toLocaleString('es', { month: 'long' }))
-
+  const [currentMonth, setCurrentMonth] = useState(new Date().toLocaleString('es', { month: 'numeric' }))
   const params = new URLSearchParams(filterParams)
   const paramsMonthName = params.get('month')
   const existMonthName = !!paramsMonthName
